+++ .gitignore
@@ -1,5 +1,6 @@
 Whisper/*
-__pycache__/*
+*__pycache__*
 *.mp4
 .env
-Recordings
+recordings
+*.mp3
