+++ README.md
@@ -11,16 +11,18 @@ python -m venv .
 pip install -r requirements.txt
 ```
 ```
-python speech2text.py
+cd src/
+python main.py
+
 ```
 
-# Setup Communications on Windows
+# Setup Communications on Windows and Linux
 
-To use this app to record minutes of an online meeting, in windows, enable the stereo mix option in advanced sound settings. Once enabled, ensure the input device is set to stereo mix. After this in advanced sound->microphone->Properties->Listen tab set listen to this device. 
+To use this app to record minutes of an online meeting, 
+- In windows, enable the stereo mix option in advanced sound settings. Once enabled, ensure the input device is set to stereo mix. After this in  
+    advanced sound->microphone->Properties->Listen tab set listen to this device. Please note this is still not perfect, you will here an echo of your voice through the speakers, the goal is to minimise this echo to acceptable levels. Contributions to this are welcome.
 
-Reduce the microphone level in the levels tab in microphone properties to minimise the echo through the speakers.
-
-Please note this is still not perfect, you will here an echo of your voice through the speakers, the goal is to minimise this echo to acceptable levels.
+- In Linux, Make sure microphone and speaker are set properly in sound settings.
