+++ requirements.txt
@@ -1,7 +1,7 @@
 docx==0.2.4
 google_api_python_client==2.110.0
 google_auth_oauthlib==1.1.0
-openai==1.3.7
+openai==1.3.8
 openai_whisper==20231117
 protobuf==4.25.1
 PyAudio==0.2.14
@@ -9,3 +9,4 @@ pycaw==20230407
 pydub==0.25.1
 python-dotenv==1.0.0
 python_docx==1.1.0
+pygame==2.5.2
