+++ token.json
@@ -1,1 +1,1 @@
-{"token": "ya29.a0AfB_byBYBgquvRFoC74vBxMhAE80pOxOYbn6ZluWMQdQGxM1y2knLTigD3JRBGmFFd4iLdNdoQQ39qAjkII095bRfXwmWFTD5T-mUd22zBzryU6KU3jQEo-eyJ6odwBQjp4d-wwvXXzht2z9AmhAdFA3t4xG9n11d-X18waCgYKAU8SARASFQHGX2MinZ01MNdMMCQZzFsg9kc6WA0173", "refresh_token": "1//0gqRT9sZCmFStCgYIARAAGBASNwF-L9Ird41SLSquyz8por7vCTjc_w2GePu-9Lohkb0TINxFwnJY7Q57D6D77iZtV9tJ1paPYDQ", "token_uri": "https://oauth2.googleapis.com/token", "client_id": "796241027558-5oe7phmkfljm7upi9n9h5nbl45l24i1u.apps.googleusercontent.com", "client_secret": "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", "scopes": ["https://www.googleapis.com/auth/drive"], "universe_domain": "googleapis.com", "expiry": "2023-12-08T13:00:49.156009Z"}
+{"token": "ya29.a0AfB_byBMsOw5D2sKyZNLQ2jKY5H9eSU2EIa6m7T0_YuGLXMZ6a1b2JIbsyVVGi9SqjPQW7HMT9lwrRCZ95zv_eWy8lefJtqMlo0j9xupZ5rBQ6-sFQbuu84QiassgvCOTTTmTp8WJcZjulo4otpSlC_elqGsqrTK2ZlGaCgYKATcSARASFQHGX2MirrGgz26c2WMGIhJhTwogzg0171", "refresh_token": "1//0gk5f5P5E6Tl0CgYIARAAGBASNwF-L9IrBslBVesv1bHxchpdXXueqLUm2b7WWPVGwFDwqyf95cvJqwRr5rkds6Txbtbz0saU7wM", "token_uri": "https://oauth2.googleapis.com/token", "client_id": "796241027558-5oe7phmkfljm7upi9n9h5nbl45l24i1u.apps.googleusercontent.com", "client_secret": "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", "scopes": ["https://www.googleapis.com/auth/drive"], "universe_domain": "googleapis.com", "expiry": "2023-12-10T15:52:44.902741Z"}
